structuring units, re-exports, docs
