+++ Cargo.toml
@@ -12,6 +12,7 @@ path = "src/main.rs"
 
 [dependencies]
 image = "0.23"
+imageproc = "0.20.0"
 glob = "0.3"
 dont_disappear = "3.0.1"
 path-absolutize = "1.0.2"
