+++ README.md
@@ -1,11 +1,48 @@
 # print-prep
 
-Command line tool for preparing photos for printing, and other bulk image transformations.
+Command line tool for preparing photos for printing, and other bulk image operations.
 
 * **[Download binaries](https://github.com/mlange-42/print-prep/releases/)**
 * [Sources on GitHub](https://github.com/mlange-42/print-prep)
 
 _This project is in a very early stage!_
+
+**Content**
+* [Features](#features)
+* [Installation](#installation)
+* [Getting started](#getting-started)
+* [Examples](#examples)
+* [Commands](#commands)
+* [Library / crate](#library--crate)
+* [License](#license)
+
+## Features
+
+* [TODO]
+
+## Installation
+
+* Download the [latest binaries](https://github.com/mlange-42/print-prep/releases/).
+* Unzip somewhere with write privileges (only required for running examples in place).
+
+## Getting started
+
+* See section [Examples](#examples) below, and try the batch files in sub-directory [/cmd_examples](https://github.com/mlange-42/print-prep/tree/master/cmd_examples). 
+* Run `pprep --help` to view the full list of options.
+
+## Examples
+
+[TODO]
+
+## Commands
+
+### Scale
+
+[TODO]
+
+### List
+
+[TODO]
 
 ## Library / crate
