+++ src/units/mod.rs
@@ -1,4 +1,15 @@
 //! Measures and units
 
 pub mod color;
-pub mod length;
+mod length;
+mod scale;
+mod size;
+
+pub use length::Length;
+pub use length::LengthUnit;
+pub use length::ToLength;
+
+pub use size::Size;
+
+pub use scale::Scale;
+pub use scale::ScaleMode;
