+++ src/util/mod.rs
@@ -1,7 +1,8 @@
 //! Utilities
 
-pub mod image;
-pub mod path;
+mod image;
+mod path;
 
+pub use self::image::ImageFormatError;
 pub use self::image::ImageUtil;
 pub use self::path::PathUtil;
